refactor auth example to import zod schemas from mutations
