+++ examples/auth/app/auth/components/LoginForm.tsx
@@ -1,8 +1,7 @@
 import {Link, useMutation, AuthenticationError} from "blitz"
 import {LabeledTextField} from "app/components/LabeledTextField"
 import {Form, FORM_ERROR} from "app/components/Form"
-import login from "app/auth/mutations/login"
-import {LoginInput} from "app/auth/validations"
+import login, {LoginInput} from "app/auth/mutations/login"
 
 type LoginFormProps = {
   onSuccess?: () => void
