+++ examples/auth/app/auth/mutations/login.ts
@@ -1,6 +1,12 @@
 import {Ctx} from "blitz"
 import {authenticateUser} from "app/auth/auth-utils"
-import {LoginInput, LoginInputType} from "../validations"
+import * as z from "zod"
+
+export const LoginInput = z.object({
+  email: z.string().email(),
+  password: z.string(),
+})
+export type LoginInputType = z.infer<typeof LoginInput>
 
 export default async function login(input: LoginInputType, {session}: Ctx) {
   // This throws an error if input is invalid
